tests of mocks not working
