+++ tests/test_network.py
@@ -1,0 +1,50 @@
+import pytest
+from unittest.mock import patch
+from unittest.mock import MagicMock, Mock
+import sys
+
+# mock Signal and Pin from Micropython and not installed
+mock_machine = MagicMock()
+sys.modules["machine"] = mock_machine
+
+
+# mock network from Micropython and not installed
+mock_network = MagicMock()
+sys.modules["network"] = mock_network
+
+# mock usockect from Micropython and not installed
+mock_usocket = MagicMock()
+sys.modules["usocket"] = mock_usocket
+
+
+from main import create_url
+
+
+# patch global variables to isolate test case
+@patch("main.WEBHOOK_KEY", "testWebhookKey")
+@patch("main.WEBHOOK_EVENT", "testWebhookEvent")
+def test_create_url():
+    action = "alarm unset"
+    assert (
+        create_url(action)
+        == "https://maker.ifttt.com/trigger/testWebhookEvent/with/key/testWebhookKey?value1=alarm unset"
+    )
+
+
+# check webook send functions
+# mock send_webhook with fake_send_webhook
+def fake_send_webhook(url):
+    full_url = "GET / HTTP/1.1\r\nHost: {}\r\n\r\n".format(url).encode()
+    print("webhook sent\n{}".format(full_url))
+    return full_url
+
+
+@patch("main.send_webhook", side_effect=fake_send_webhook)
+def test_send_webhook(send_webhook):
+    url = create_url("test action")
+    full_url = send_webhook(url)
+    assert type(full_url) == bytes
+    assert (
+        full_url
+        == b"GET / HTTP/1.1\r\nHost: https://maker.ifttt.com/trigger/alarm_activated/with/key/{IFTTT webhook key}?value1=test action\r\n\r\n"
+    )
